Wall_detection/requirements.txt
Django==5.2.1
torch==2.2.2+cpu
torchvision==0.17.2+cpu
Pillow
numpy
opencv-python
matplotlib